+++ programs/drift/src/controller/funding.rs
@@ -271,6 +271,7 @@ pub fn update_funding_rate(
             .safe_add(funding_rate_short)?;
 
         market.amm.last_funding_rate = funding_rate;
+        market.amm.last_funding_oracle_twap = oracle_price_twap;
         market.amm.last_funding_rate_long = funding_rate_long.cast()?;
         market.amm.last_funding_rate_short = funding_rate_short.cast()?;
         market.amm.last_24h_avg_funding_rate = calculate_new_twap(