+++ programs/drift/src/state/perp_market.rs
@@ -10,17 +10,17 @@ use crate::error::{DriftResult, ErrorCode};
 use crate::math::amm;
 use crate::math::casting::Cast;
 #[cfg(test)]
-use crate::math::constants::{
-    AMM_RESERVE_PRECISION, MAX_CONCENTRATION_COEFFICIENT, PRICE_PRECISION_I64,
-};
+use crate::math::constants::{AMM_RESERVE_PRECISION, MAX_CONCENTRATION_COEFFICIENT};
 use crate::math::constants::{
     AMM_RESERVE_PRECISION_I128, AMM_TO_QUOTE_PRECISION_RATIO, BID_ASK_SPREAD_PRECISION,
     BID_ASK_SPREAD_PRECISION_I128, BID_ASK_SPREAD_PRECISION_U128,
-    DEFAULT_REVENUE_SINCE_LAST_FUNDING_SPREAD_RETREAT, LIQUIDATION_FEE_PRECISION,
+    DEFAULT_REVENUE_SINCE_LAST_FUNDING_SPREAD_RETREAT, FUNDING_RATE_BUFFER_I128,
+    FUNDING_RATE_OFFSET_PERCENTAGE, LIQUIDATION_FEE_PRECISION,
     LIQUIDATION_FEE_TO_MARGIN_PRECISION_RATIO, LP_FEE_SLICE_DENOMINATOR, LP_FEE_SLICE_NUMERATOR,
     MARGIN_PRECISION, MARGIN_PRECISION_U128, MAX_LIQUIDATION_MULTIPLIER, PEG_PRECISION,
     PERCENTAGE_PRECISION, PERCENTAGE_PRECISION_I128, PERCENTAGE_PRECISION_I64,
-    PERCENTAGE_PRECISION_U64, PRICE_PRECISION, SPOT_WEIGHT_PRECISION, TWENTY_FOUR_HOUR,
+    PERCENTAGE_PRECISION_U64, PRICE_PRECISION, PRICE_PRECISION_I128, PRICE_PRECISION_I64,
+    SPOT_WEIGHT_PRECISION, TWENTY_FOUR_HOUR,
 };
 use crate::math::helpers::get_proportion_i128;
 use crate::math::margin::{
@@ -253,7 +253,9 @@ pub struct PerpMarket {
     pub high_leverage_margin_ratio_maintenance: u16,
     pub protected_maker_limit_price_divisor: u8,
     pub protected_maker_dynamic_divisor: u8,
-    pub padding: [u8; 36],
+    pub padding1: u32,
+    pub last_fill_price: u64,
+    pub padding: [u8; 24],
 }
 
 impl Default for PerpMarket {
@@ -295,7 +297,9 @@ impl Default for PerpMarket {
             high_leverage_margin_ratio_maintenance: 0,
             protected_maker_limit_price_divisor: 0,
             protected_maker_dynamic_divisor: 0,
-            padding: [0; 36],
+            padding1: 0,
+            last_fill_price: 0,
+            padding: [0; 24],
         }
     }
 }
@@ -750,6 +754,109 @@ impl PerpMarket {
         }
     }
 
+    pub fn get_trigger_price(
+        &self,
+        oracle_price: i64,
+        now: i64,
+        use_median_price: bool,
+    ) -> DriftResult<u64> {
+        if !use_median_price {
+            return oracle_price.cast::<u64>();
+        }
+
+        let last_fill_price = self.last_fill_price;
+
+        let mark_price_5min_twap = self.amm.last_mark_price_twap;
+        let last_oracle_price_twap_5min =
+            self.amm.historical_oracle_data.last_oracle_price_twap_5min;
+
+        let basis_5min = mark_price_5min_twap
+            .cast::<i64>()?
+            .safe_sub(last_oracle_price_twap_5min)?;
+
+        let oracle_plus_basis_5min = oracle_price.safe_add(basis_5min)?.cast::<u64>()?;
+
+        let last_funding_basis = self.get_last_funding_basis(oracle_price, now)?;
+
+        let oracle_plus_funding_basis = oracle_price.safe_add(last_funding_basis)?.cast::<u64>()?;
+
+        let median_price = if last_fill_price > 0 {
+            let mut prices = [
+                last_fill_price,
+                oracle_plus_funding_basis,
+                oracle_plus_basis_5min,
+            ];
+            prices.sort_unstable();
+
+            prices[1]
+        } else {
+            let mut prices = [
+                oracle_price.unsigned_abs(),
+                oracle_plus_funding_basis,
+                oracle_plus_basis_5min,
+            ];
+            prices.sort_unstable();
+
+            prices[1]
+        };
+
+        self.clamp_trigger_price(oracle_price.unsigned_abs(), median_price)
+    }
+
+    #[inline(always)]
+    fn get_last_funding_basis(&self, oracle_price: i64, now: i64) -> DriftResult<i64> {
+        if self.amm.last_funding_oracle_twap > 0 {
+            let last_funding_rate = self
+                .amm
+                .last_funding_rate
+                .cast::<i128>()?
+                .safe_mul(PRICE_PRECISION_I128)?
+                .safe_div(self.amm.last_funding_oracle_twap.cast::<i128>()?)?
+                .safe_mul(24)?;
+            let last_funding_rate_pre_adj =
+                last_funding_rate.safe_sub(FUNDING_RATE_OFFSET_PERCENTAGE as i128)?;
+
+            let time_left_until_funding_update = now
+                .safe_sub(self.amm.last_funding_rate_ts)?
+                .min(self.amm.funding_period);
+
+            let last_funding_basis = oracle_price
+                .cast::<i128>()?
+                .safe_mul(last_funding_rate_pre_adj)?
+                .safe_div(PERCENTAGE_PRECISION_I128)?
+                .safe_mul(
+                    self.amm
+                        .funding_period
+                        .safe_sub(time_left_until_funding_update)?
+                        .cast::<i128>()?,
+                )?
+                .safe_div(self.amm.funding_period.cast::<i128>()?)?
+                / FUNDING_RATE_BUFFER_I128;
+
+            last_funding_basis.cast::<i64>()
+        } else {
+            Ok(0)
+        }
+    }
+
+    #[inline(always)]
+    fn clamp_trigger_price(&self, oracle_price: u64, median_price: u64) -> DriftResult<u64> {
+        let max_bps_diff = if matches!(self.contract_tier, ContractTier::A | ContractTier::B) {
+            500 // 20 BPS
+        } else if matches!(self.contract_tier, ContractTier::C) {
+            100 // 100 BPS
+        } else {
+            40 // 250 BPS
+        };
+        let max_oracle_diff = oracle_price / max_bps_diff;
+
+        Ok(median_price.clamp(
+            oracle_price.safe_sub(max_oracle_diff)?,
+            oracle_price.safe_add(max_oracle_diff)?,
+        ))
+    }
+
+    #[inline(always)]
     pub fn get_mm_oracle_price_data(
         &self,
         oracle_price_data: OraclePriceData,
@@ -952,10 +1059,10 @@ pub struct AMM {
     /// precision: AMM_RESERVE_PRECISION
     pub user_lp_shares: u128,
     /// last funding rate in this perp market (unit is quote per base)
-    /// precision: QUOTE_PRECISION
+    /// precision: FUNDING_RATE_PRECISION
     pub last_funding_rate: i64,
     /// last funding rate for longs in this perp market (unit is quote per base)
-    /// precision: QUOTE_PRECISION
+    /// precision: FUNDING_RATE_PRECISION
     pub last_funding_rate_long: i64,
     /// last funding rate for shorts in this perp market (unit is quote per base)
     /// precision: QUOTE_PRECISION
@@ -1096,7 +1203,8 @@ pub struct AMM {
     pub reference_price_offset: i32,
     /// signed scale amm_spread similar to fee_adjustment logic (-100 = 0, 100 = double)
     pub amm_inventory_spread_adjustment: i8,
-    pub padding: [u8; 11],
+    pub padding: [u8; 3],
+    pub last_funding_oracle_twap: i64,
 }
 
 impl Default for AMM {
@@ -1186,7 +1294,8 @@ impl Default for AMM {
             quote_asset_amount_with_unsettled_lp: 0,
             reference_price_offset: 0,
             amm_inventory_spread_adjustment: 0,
-            padding: [0; 11],
+            padding: [0; 3],
+            last_funding_oracle_twap: 0,
         }
     }
 }