+++ sdk/src/dlob/DLOB.ts
@@ -11,6 +11,7 @@ import { decodeName } from '../userName';
 import { DLOBNode, DLOBNodeType, TriggerOrderNode } from './DLOBNode';
 import { DriftClient } from '../driftClient';
 import {
+	calculateOrderBaseAssetAmount,
 	getLimitPrice,
 	isOrderExpired,
 	isRestingLimitOrder,
@@ -180,7 +181,26 @@ export class DLOB {
 			const protectedMaker = isUserProtectedMaker(userAccount);
 
 			for (const order of userAccount.orders) {
-				this.insertOrder(order, userAccountPubkeyString, slot, protectedMaker);
+				let baseAssetAmount = order.baseAssetAmount;
+				if (order.reduceOnly) {
+					const existingBaseAmount =
+						userAccount.perpPositions.find(
+							(pos) =>
+								pos.marketIndex === order.marketIndex && pos.openOrders > 0
+						)?.baseAssetAmount || ZERO;
+					baseAssetAmount = calculateOrderBaseAssetAmount(
+						order,
+						existingBaseAmount
+					);
+				}
+
+				this.insertOrder(
+					order,
+					userAccountPubkeyString,
+					slot,
+					protectedMaker,
+					baseAssetAmount
+				);
 			}
 		}
 
@@ -193,6 +213,7 @@ export class DLOB {
 		userAccount: string,
 		slot: number,
 		isUserProtectedMaker: boolean,
+		baseAssetAmount: BN,
 		onInsert?: OrderBookCallback
 	): void {
 		if (!isVariant(order.status, 'open')) {
@@ -220,7 +241,8 @@ export class DLOB {
 			marketType,
 			userAccount,
 			isUserProtectedMaker,
-			this.protectedMakerParamsMap[marketType].get(order.marketIndex)
+			this.protectedMakerParamsMap[marketType].get(order.marketIndex),
+			baseAssetAmount
 		);
 
 		if (onInsert) {
@@ -232,6 +254,7 @@ export class DLOB {
 		order: Order,
 		userAccount: string,
 		isUserProtectedMaker: boolean,
+		baseAssetAmount?: BN,
 		onInsert?: OrderBookCallback
 	): void {
 		const marketType = getVariant(order.marketType) as MarketTypeStr;
@@ -251,7 +274,8 @@ export class DLOB {
 				marketType,
 				userAccount,
 				isUserProtectedMaker,
-				this.protectedMakerParamsMap[marketType].get(order.marketIndex)
+				this.protectedMakerParamsMap[marketType].get(order.marketIndex),
+				baseAssetAmount
 			);
 		if (onInsert) {
 			onInsert();
@@ -1558,7 +1582,7 @@ export class DLOB {
 	public findNodesToTrigger(
 		marketIndex: number,
 		slot: number,
-		oraclePrice: BN,
+		triggerPrice: BN,
 		marketType: MarketType,
 		stateAccount: StateAccount
 	): NodeToTrigger[] {
@@ -1575,7 +1599,7 @@ export class DLOB {
 			: undefined;
 		if (triggerAboveList) {
 			for (const node of triggerAboveList.getGenerator()) {
-				if (oraclePrice.gt(node.order.triggerPrice)) {
+				if (triggerPrice.gt(node.order.triggerPrice)) {
 					nodesToTrigger.push({
 						node: node,
 					});
@@ -1590,7 +1614,7 @@ export class DLOB {
 			: undefined;
 		if (triggerBelowList) {
 			for (const node of triggerBelowList.getGenerator()) {
-				if (oraclePrice.lt(node.order.triggerPrice)) {
+				if (triggerPrice.lt(node.order.triggerPrice)) {
 					nodesToTrigger.push({
 						node: node,
 					});