+++ sdk/src/driftClient.ts
@@ -64,6 +64,7 @@ import {
 	ProtectedMakerModeConfig,
 	SignedMsgOrderParamsDelegateMessage,
 	TokenProgramFlag,
+	PostOnlyParams,
 } from './types';
 import driftIDL from './idl/drift.json';
 
@@ -123,6 +124,8 @@ import { TxSender, TxSigAndSlot } from './tx/types';
 import {
 	BASE_PRECISION,
 	GOV_SPOT_MARKET_INDEX,
+	ONE,
+	PERCENTAGE_PRECISION,
 	PRICE_PRECISION,
 	QUOTE_PRECISION,
 	QUOTE_SPOT_MARKET_INDEX,
@@ -444,7 +447,9 @@ export class DriftClient {
 					resubTimeoutMs: config.accountSubscription?.resubTimeoutMs,
 					logResubMessages: config.accountSubscription?.logResubMessages,
 				},
-				config.accountSubscription?.commitment
+				config.accountSubscription?.commitment,
+				config.accountSubscription?.perpMarketAccountSubscriber,
+				config.accountSubscription?.oracleAccountSubscriber
 			);
 		}
 		this.eventEmitter = this.accountSubscriber.eventEmitter;
@@ -4163,25 +4168,20 @@ export class DriftClient {
 	public async getUpdateAMMsIx(
 		marketIndexes: number[]
 	): Promise<TransactionInstruction> {
-		for (let i = marketIndexes.length; i < 5; i++) {
-			marketIndexes.push(100);
-		}
 		const marketAccountInfos = [];
 		const oracleAccountInfos = [];
 		for (const marketIndex of marketIndexes) {
-			if (marketIndex !== 100) {
-				const market = this.getPerpMarketAccount(marketIndex);
-				marketAccountInfos.push({
-					pubkey: market.pubkey,
-					isWritable: true,
-					isSigner: false,
-				});
-				oracleAccountInfos.push({
-					pubkey: market.amm.oracle,
-					isWritable: false,
-					isSigner: false,
-				});
-			}
+			const market = this.getPerpMarketAccount(marketIndex);
+			marketAccountInfos.push({
+				pubkey: market.pubkey,
+				isWritable: true,
+				isSigner: false,
+			});
+			oracleAccountInfos.push({
+				pubkey: market.amm.oracle,
+				isWritable: false,
+				isSigner: false,
+			});
 		}
 		const remainingAccounts = oracleAccountInfos.concat(marketAccountInfos);
 
@@ -4354,14 +4354,24 @@ export class DriftClient {
 		});
 	}
 
+	/**
+	 * Sends a transaction to cancel the provided order ids.
+	 *
+	 * @param orderIds - The order ids to cancel.
+	 * @param txParams - The transaction parameters.
+	 * @param subAccountId - The sub account id to cancel the orders for.
+	 * @param user - The user to cancel the orders for. If provided, it will be prioritized over the subAccountId.
+	 * @returns The transaction signature.
+	 */
 	public async cancelOrdersByIds(
 		orderIds?: number[],
 		txParams?: TxParams,
-		subAccountId?: number
+		subAccountId?: number,
+		user?: User
 	): Promise<TransactionSignature> {
 		const { txSig } = await this.sendTransaction(
 			await this.buildTransaction(
-				await this.getCancelOrdersByIdsIx(orderIds, subAccountId),
+				await this.getCancelOrdersByIdsIx(orderIds, subAccountId, user),
 				txParams
 			),
 			[],
@@ -4370,21 +4380,34 @@ export class DriftClient {
 		return txSig;
 	}
 
+	/**
+	 * Returns the transaction instruction to cancel the provided order ids.
+	 *
+	 * @param orderIds - The order ids to cancel.
+	 * @param subAccountId - The sub account id to cancel the orders for.
+	 * @param user - The user to cancel the orders for. If provided, it will be prioritized over the subAccountId.
+	 * @returns The transaction instruction to cancel the orders.
+	 */
 	public async getCancelOrdersByIdsIx(
 		orderIds?: number[],
-		subAccountId?: number
+		subAccountId?: number,
+		user?: User
 	): Promise<TransactionInstruction> {
-		const user = await this.getUserAccountPublicKey(subAccountId);
+		const userAccountPubKey =
+			user?.userAccountPublicKey ??
+			(await this.getUserAccountPublicKey(subAccountId));
+		const userAccount =
+			user?.getUserAccount() ?? this.getUserAccount(subAccountId);
 
 		const remainingAccounts = this.getRemainingAccounts({
-			userAccounts: [this.getUserAccount(subAccountId)],
+			userAccounts: [userAccount],
 			useMarketLastSlotCache: true,
 		});
 
 		return await this.program.instruction.cancelOrdersByIds(orderIds, {
 			accounts: {
 				state: await this.getStatePublicKey(),
-				user,
+				user: userAccountPubKey,
 				authority: this.wallet.publicKey,
 			},
 			remainingAccounts,
@@ -6947,7 +6970,7 @@ export class DriftClient {
 			auctionStartPrice?: BN;
 			auctionEndPrice?: BN;
 			reduceOnly?: boolean;
-			postOnly?: boolean;
+			postOnly?: PostOnlyParams;
 			bitFlags?: number;
 			maxTs?: BN;
 			policy?: number;
@@ -6966,6 +6989,12 @@ export class DriftClient {
 		return txSig;
 	}
 
+	/**
+	 * @param orderParams: The parameters for the order to modify.
+	 * @param subAccountId: Optional - The subaccount ID of the user to modify the order for.
+	 * @param userPublicKey: Optional - The public key of the user to modify the order for. This takes precedence over subAccountId.
+	 * @returns
+	 */
 	public async getModifyOrderIx(
 		{
 			orderId,
@@ -6995,14 +7024,16 @@ export class DriftClient {
 			auctionStartPrice?: BN;
 			auctionEndPrice?: BN;
 			reduceOnly?: boolean;
-			postOnly?: boolean;
+			postOnly?: PostOnlyParams;
 			bitFlags?: number;
 			maxTs?: BN;
 			policy?: number;
 		},
-		subAccountId?: number
+		subAccountId?: number,
+		userPublicKey?: PublicKey
 	): Promise<TransactionInstruction> {
-		const user = await this.getUserAccountPublicKey(subAccountId);
+		const user =
+			userPublicKey ?? (await this.getUserAccountPublicKey(subAccountId));
 
 		const remainingAccounts = this.getRemainingAccounts({
 			userAccounts: [this.getUserAccount(subAccountId)],
@@ -7068,7 +7099,7 @@ export class DriftClient {
 			auctionStartPrice?: BN;
 			auctionEndPrice?: BN;
 			reduceOnly?: boolean;
-			postOnly?: boolean;
+			postOnly?: PostOnlyParams;
 			bitFlags?: number;
 			policy?: ModifyOrderPolicy;
 			maxTs?: BN;
@@ -7116,7 +7147,7 @@ export class DriftClient {
 			auctionStartPrice?: BN;
 			auctionEndPrice?: BN;
 			reduceOnly?: boolean;
-			postOnly?: boolean;
+			postOnly?: PostOnlyParams;
 			bitFlags?: number;
 			policy?: ModifyOrderPolicy;
 			maxTs?: BN;
@@ -8485,15 +8516,25 @@ export class DriftClient {
 	}
 
 	public getOracleDataForPerpMarket(marketIndex: number): OraclePriceData {
-		return this.accountSubscriber.getOraclePriceDataAndSlotForPerpMarket(
-			marketIndex
-		).data;
+		const perpMarket = this.getPerpMarketAccount(marketIndex);
+		const isMMOracleActive = !perpMarket.amm.mmOracleSlot.eq(ZERO);
+		return {
+			...this.accountSubscriber.getOraclePriceDataAndSlotForPerpMarket(
+				marketIndex
+			).data,
+			isMMOracleActive,
+		};
 	}
 
 	public getMMOracleDataForPerpMarket(marketIndex: number): OraclePriceData {
 		const perpMarket = this.getPerpMarketAccount(marketIndex);
 		const oracleData = this.getOracleDataForPerpMarket(marketIndex);
 		const stateAccountAndSlot = this.accountSubscriber.getStateAccountAndSlot();
+		const pctDiff = perpMarket.amm.mmOraclePrice
+			.sub(oracleData.price)
+			.abs()
+			.mul(PERCENTAGE_PRECISION)
+			.div(BN.max(oracleData.price, ONE));
 		if (
 			!isOracleValid(
 				perpMarket,
@@ -8502,7 +8543,8 @@ export class DriftClient {
 				stateAccountAndSlot.slot
 			) ||
 			perpMarket.amm.mmOraclePrice.eq(ZERO) ||
-			perpMarket.amm.mmOracleSlot < oracleData.slot
+			perpMarket.amm.mmOracleSlot < oracleData.slot ||
+			pctDiff.gt(PERCENTAGE_PRECISION.divn(100)) // 1% threshold
 		) {
 			return { ...oracleData, fetchedWithMMOracle: true };
 		} else {
@@ -8517,6 +8559,7 @@ export class DriftClient {
 				confidence: conf,
 				hasSufficientNumberOfDataPoints: true,
 				fetchedWithMMOracle: true,
+				isMMOracleActive: oracleData.isMMOracleActive,
 			};
 		}
 	}
@@ -9839,17 +9882,20 @@ export class DriftClient {
 			  })
 			: undefined;
 
-		const ix = await this.program.instruction.disableUserHighLeverageMode({
-			accounts: {
-				state: await this.getStatePublicKey(),
-				user,
-				authority: this.wallet.publicKey,
-				highLeverageModeConfig: getHighLeverageModeConfigPublicKey(
-					this.program.programId
-				),
-			},
-			remainingAccounts,
-		});
+		const ix = await this.program.instruction.disableUserHighLeverageMode(
+			false,
+			{
+				accounts: {
+					state: await this.getStatePublicKey(),
+					user,
+					authority: this.wallet.publicKey,
+					highLeverageModeConfig: getHighLeverageModeConfigPublicKey(
+						this.program.programId
+					),
+				},
+				remainingAccounts,
+			}
+		);
 
 		return ix;
 	}